VacationRental.Contracts assembly was created
