--- VacationRental.Api/ViewModels/CalendarBookingViewModel.cs
@@ -1,19 +1,0 @@
-﻿using VacationRental.Api.Models;
-
-namespace VacationRental.Api.ViewModels
-{
-    public class CalendarBookingViewModel
-    {
-        public CalendarBookingViewModel(int id)
-        {
-            Id = id;
-        }
-        
-        public int Id { get; }
-        
-        public static CalendarBookingViewModel FromBooking(Booking booking)
-        {
-            return new CalendarBookingViewModel(booking.Id);
-        }
-    }
-}
