+++ VacationRental.Contracts/Calendar/CalendarBookingViewModel.cs
@@ -1,0 +1,12 @@
+﻿namespace VacationRental.Contracts.Calendar
+{
+    public class CalendarBookingViewModel
+    {
+        public CalendarBookingViewModel(int id)
+        {
+            Id = id;
+        }
+        
+        public int Id { get; }
+    }
+}
