+++ VacationRental.Contracts/Calendar/CalendarDateViewModel.cs
@@ -1,0 +1,14 @@
+﻿namespace VacationRental.Contracts.Calendar
+{
+    public class CalendarDateViewModel
+    {
+        public CalendarDateViewModel(DateTime date, List<CalendarBookingViewModel> bookings)
+        {
+            Date = date;
+            Bookings = bookings ?? throw new ArgumentNullException(nameof(bookings));
+        }
+        
+        public DateTime Date { get; }
+        public List<CalendarBookingViewModel> Bookings { get; }
+    }
+}
